+++ lib/esv_bible.dart
@@ -1,3 +1,7 @@
 library esv_bible;
 
+export 'src/domain/entities/passage_audio.dart';
+export 'src/domain/entities/passage_text.dart';
+export 'src/domain/entities/passage_search.dart';
+export 'src/domain/entities/passage_html.dart';
 export 'src/passage.dart';