+++ test/src/data/data_sources/esv_remote_data_source_test.dart
@@ -1,7 +1,9 @@
-import 'package:esv_bible/src/data/data_sources/esv_remote_data_source.dart';
 import 'package:test/test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:http/http.dart' as http;
+
+import 'package:esv_bible/src/data/data_sources/esv_remote_data_source_implementation.dart';
+
 import '../../../test_data.dart';
 
 class MockHttpClient extends Mock implements http.Client {}